# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TRIMESTRA API",
        "description": "API Documentation for TRIMESTRA - Pregnancy Nutrition App",
        "contact": {
            "email": "support@trimestra.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://trimestra-backend.test/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/diary": {
            "get": {
                "tags": [
                    "Diary"
                ],
                "summary": "Get diary entries for a date",
                "operationId": "d625fec3dd2e9dc1223202f24be79ca8",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date in Y-m-d format",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diary entries retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "entries": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/DiaryEntry"
                                                    }
                                                },
                                                "summary": {
                                                    "type": "object"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Diary entries retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Diary"
                ],
                "summary": "Add diary entry",
                "operationId": "d8f988f91e5341334dcdb684bb9a5121",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "meal_type"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "meal_type": {
                                        "type": "string",
                                        "enum": [
                                            "breakfast",
                                            "lunch",
                                            "dinner",
                                            "snack",
                                            "exercise"
                                        ]
                                    },
                                    "food_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "recipe_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.5
                                    },
                                    "exercise_type": {
                                        "type": "string",
                                        "example": "walking"
                                    },
                                    "exercise_duration": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Diary entry added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DiaryEntry"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Diary entry added successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/diary/summary": {
            "get": {
                "tags": [
                    "Diary"
                ],
                "summary": "Get daily summary",
                "operationId": "122a8a1d54461e3081b6ce19781ec636",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date in Y-m-d format",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daily summary retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "total_calories_intake": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "total_calories_burned": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "total_protein": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "total_carbs": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "total_fat": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "total_fiber": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "target_calories": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daily summary retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DiaryEntry": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "meal_type": {
                        "type": "string",
                        "enum": [
                            "breakfast",
                            "lunch",
                            "dinner",
                            "snack",
                            "exercise"
                        ]
                    },
                    "food_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "recipe_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "number",
                        "format": "float"
                    },
                    "calories": {
                        "type": "number",
                        "format": "float"
                    },
                    "protein": {
                        "type": "number",
                        "format": "float"
                    },
                    "carbs": {
                        "type": "number",
                        "format": "float"
                    },
                    "fat": {
                        "type": "number",
                        "format": "float"
                    },
                    "fiber": {
                        "type": "number",
                        "format": "float"
                    },
                    "exercise_type": {
                        "type": "string",
                        "nullable": true
                    },
                    "exercise_duration": {
                        "type": "integer",
                        "nullable": true
                    },
                    "calories_burned": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "device_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}